Created fragments with view models for entering personal data and applying loan request, added network DTOs for loan request and response, created use case

+++ app/src/main/java/com/girrafeecstud/final_loan_app_zalessky/data/network/loan/dto/LoanConditionsResponse.kt
@@ -4,7 +4,7 @@ import com.google.gson.annotations.SerializedName
 
 data class LoanConditionsResponse(
     @SerializedName("maxAmount")
-    val maxLoanAmount: Int,
+    val maxLoanAmount: Double,
     @SerializedName("percent")
     val loanPercent: Double,
     @SerializedName("period")
