+++ app/src/main/java/com/girrafeecstud/final_loan_app_zalessky/domain/entities/LoanConditions.kt
@@ -1,7 +1,7 @@
 package com.girrafeecstud.final_loan_app_zalessky.domain.entities
 
 data class LoanConditions(
-    val maxAmount: Int,
+    val maxAmount: Double,
     val percent: Double,
     val period: Int
 )
